Move graph to its own file.
