+++ toonz/sources/toonz/CMakeLists.txt
@@ -121,6 +121,7 @@ set(MOC_HEADERS
     quicktoolbar.h
 	xdtsimportpopup.h
     motionpathpanel.h
+    graphwidget.h
     ../stopmotion/stopmotion.h
     ../stopmotion/stopmotioncontroller.h
     ../stopmotion/webcam.h
@@ -351,6 +352,7 @@ set(SOURCES
     ObjectTracker.cpp
     predict3d.cpp
     motionpathpanel.cpp
+    graphwidget.cpp
     ../stopmotion/stopmotion.cpp
     ../stopmotion/stopmotioncontroller.cpp
     ../stopmotion/webcam.cpp
