+++ toonz/sources/toonz/motionpathpanel.h
@@ -22,6 +22,7 @@ class QToolBar;
 class QSlider;
 class QComboBox;
 class TThickPoint;
+class GraphWidget;
 
 
 //=============================================================================
@@ -68,45 +69,6 @@ protected:
 };
 
 //=============================================================================
-// GraphArea
-//-----------------------------------------------------------------------------
-
-class GraphArea : public QWidget
-{
-    Q_OBJECT
-
-    TStroke *m_stroke;
-    double m_maxXValue;
-    double m_maxYValue;
-    int m_selectedControlPoint = -1;
-public:
-    explicit GraphArea(QWidget* parent = nullptr);
-
-    QSize minimumSizeHint() const override;
-    QSize sizeHint() const override;
-    void setStroke(TStroke *stroke) { m_stroke = stroke; }
-    void clearStroke() { m_stroke = 0; }
-    void setMaxXValue(int x) { m_maxXValue = x; }
-    void setMaxYValue(int y) { m_maxYValue = y; }
-    //public slots:
-    //void setAntialiased(bool antialiased);
-
-protected:
-    void paintEvent(QPaintEvent* event) override;
-    void mousePressEvent(QMouseEvent* event) override;
-    void mouseMoveEvent(QMouseEvent* event) override;
-    void mouseReleaseEvent(QMouseEvent* event) override;
-    QPointF convertPointToInvertedLocal(TThickPoint point);
-    TThickPoint convertInvertedLocalToPoint(QPointF point);
-
-private:
-    QPen pen;
-    QBrush brush;
-    bool antialiased;
-    QPixmap pixmap;
-};
-
-//=============================================================================
 // MotionPathPanel
 //-----------------------------------------------------------------------------
 
@@ -122,7 +84,7 @@ class MotionPathPanel final : public QWidget {
   QToolBar* m_toolbar;
   std::vector<MotionPathControl*> m_motionPathControls;
   TStageObjectSpline *m_currentSpline;
-  GraphArea *m_graphArea;
+  GraphWidget *m_graphArea;
 public:
   MotionPathPanel(QWidget* parent = 0);
   ~MotionPathPanel();
